refactor(skills): 将 SkillAuditReport 组件移入模块内部并修正引用路径
