+++ src/renderer/pages/settings/skills/components/SkillAuditReport.tsx
